chore: transition to esm
All scripts are now in esm format.  This forced the removal of the yafu
acceptance tests that tested that the bundles worked as expected. They
should not be needed anymore however since we now use modern bundling
tools that is expected to keep the same functionality as the original
code.

Transpiling has been removed from tests.